Add files via upload
This code is modified to use with 16 chan headstages (with or without 3-axis accelerometers) and 2 analog inputs per headstage. Additionally, the code is a little more flexible in terms of headstage port settings on the INTAN GUI when recording. Even if the 'blank' channels on a particular INTAN port were not disabled when recording the rhd files, this code will be able to extract the actual ephys chans correctly.