Add time resolution bar overlay to spectrogram
Spectrogram generating plugins (egs_*) now must return a second argument as well which is the temporal resolution of the spectrogram in seconds.

Also got rid of the thing where electro_gui uses the existence of text on the spectrogram axes to determine if a macro is in progress. This will probably cause problems - need to finish implementing a modern progress indicator/UI freeze.